requirements.txt
Django==4.2.21
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
mysqlclient==2.2.4  # hoặc pymysql==1.1.0 nếu bạn dùng pymysql
drf-spectacular==0.28.0
django-filter==25.1